Add campus test data and add loops to menus

+++ src/history.py
@@ -4,7 +4,7 @@ from data_structures import Stack
 class History:
     def __init__(self):
         self.nav = Stack()
-    def add(self,start : Building, end : Building, along : Stack, total : int):
+    def add(self,start : Building, end : Building, along : Stack, total : float):
         #starting point, end  point, buildings passed on the way in order, total time spent
         self.nav.push([start,end,along,total])
     def remove(self):
